+++ src/main/java/org/opensearch/knn/index/codec/util/KNNCodecUtil.java
@@ -17,6 +17,7 @@ import org.opensearch.knn.common.KNNConstants;
 import org.opensearch.knn.index.VectorDataType;
 import org.opensearch.knn.index.codec.KNN80Codec.KNN80BinaryDocValues;
 import org.opensearch.knn.index.engine.KNNEngine;
+import org.opensearch.knn.index.engine.VectorSearchEngine;
 import org.opensearch.knn.index.vectorvalues.KNNVectorValues;
 
 import java.io.IOException;
@@ -80,7 +81,7 @@ public class KNNCodecUtil {
     /**
      * Get Engine Files from segment with specific fieldName and engine extension
      *
-     * @param extension Engine extension comes from {@link KNNEngine#getExtension()}}
+     * @param extension Engine extension comes from {@link VectorSearchEngine#getExtension()}}
      * @param fieldName Filed for knn field
      * @param segmentInfo {@link SegmentInfo} One Segment info to use for compute.
      * @return List of engine files
@@ -114,7 +115,7 @@ public class KNNCodecUtil {
             return null;
         }
         // Only Native Engine put into indexPathMap
-        final KNNEngine knnEngine = getNativeKNNEngine(field);
+        final VectorSearchEngine knnEngine = getNativeKNNEngine(field);
         if (knnEngine == null) {
             return null;
         }
@@ -158,8 +159,8 @@ public class KNNCodecUtil {
      * @param field which field we need produce from engine
      * @return if and only if Native Engine we return specific engine, else return null
      */
-    private static KNNEngine getNativeKNNEngine(@NonNull FieldInfo field) {
-        final KNNEngine engine = FieldInfoExtractor.extractKNNEngine(field);
+    private static VectorSearchEngine getNativeKNNEngine(@NonNull FieldInfo field) {
+        final VectorSearchEngine engine = FieldInfoExtractor.extractKNNEngine(field);
         if (KNNEngine.getEnginesThatCreateCustomSegmentFiles().contains(engine)) {
             return engine;
         }