+++ src/test/java/org/opensearch/knn/index/codec/KNN80Codec/KNN80DocValuesConsumerTests.java
@@ -44,6 +44,7 @@ import org.opensearch.knn.index.SpaceType;
 import org.opensearch.knn.index.codec.KNNCodecTestUtil;
 import org.opensearch.knn.index.codec.util.KNNCodecUtil;
 import org.opensearch.knn.index.engine.KNNEngine;
+import org.opensearch.knn.index.engine.VectorSearchEngine;
 import org.opensearch.knn.indices.Model;
 import org.opensearch.knn.indices.ModelCache;
 import org.opensearch.knn.indices.ModelDao;
@@ -200,7 +201,7 @@ public class KNN80DocValuesConsumerTests extends KNNTestCase {
         int docsInSegment = 100;
         String fieldName = String.format(Locale.ROOT, "test_field%s", randomAlphaOfLength(4));
 
-        KNNEngine knnEngine = KNNEngine.NMSLIB;
+        VectorSearchEngine knnEngine = KNNEngine.NMSLIB;
         SpaceType spaceType = SpaceType.COSINESIMIL;
         int dimension = 16;
 
@@ -268,7 +269,7 @@ public class KNN80DocValuesConsumerTests extends KNNTestCase {
         int docsInSegment = 100;
         String fieldName = String.format(Locale.ROOT, "test_field%s", randomAlphaOfLength(4));
 
-        KNNEngine knnEngine = KNNEngine.NMSLIB;
+        VectorSearchEngine knnEngine = KNNEngine.NMSLIB;
         SpaceType spaceType = SpaceType.COSINESIMIL;
         int dimension = 16;
 
@@ -322,7 +323,7 @@ public class KNN80DocValuesConsumerTests extends KNNTestCase {
         int docsInSegment = 100;
         String fieldName = String.format(Locale.ROOT, "test_field%s", randomAlphaOfLength(4));
 
-        KNNEngine knnEngine = KNNEngine.FAISS;
+        VectorSearchEngine knnEngine = KNNEngine.FAISS;
         SpaceType spaceType = SpaceType.INNER_PRODUCT;
         int dimension = 16;
 
@@ -387,7 +388,7 @@ public class KNN80DocValuesConsumerTests extends KNNTestCase {
         int docsInSegment = 100;
         String fieldName = String.format(Locale.ROOT, "test_field%s", randomAlphaOfLength(4));
 
-        KNNEngine knnEngine = KNNEngine.FAISS;
+        VectorSearchEngine knnEngine = KNNEngine.FAISS;
         SpaceType spaceType = SpaceType.HAMMING;
         VectorDataType dataType = VectorDataType.BINARY;
         int dimension = 16;
@@ -453,7 +454,7 @@ public class KNN80DocValuesConsumerTests extends KNNTestCase {
 
     public void testAddKNNBinaryField_fromModel_faiss() throws IOException, ExecutionException, InterruptedException {
         // Generate a trained faiss model
-        KNNEngine knnEngine = KNNEngine.FAISS;
+        VectorSearchEngine knnEngine = KNNEngine.FAISS;
         SpaceType spaceType = SpaceType.INNER_PRODUCT;
         int dimension = 16;
         String modelId = "test-model-id";
@@ -628,7 +629,7 @@ public class KNN80DocValuesConsumerTests extends KNNTestCase {
 
     public void testAddKNNBinaryField_fromModel_faiss_with_mergeAbort() throws IOException {
         // Generate a trained faiss model
-        KNNEngine knnEngine = KNNEngine.FAISS;
+        VectorSearchEngine knnEngine = KNNEngine.FAISS;
         SpaceType spaceType = SpaceType.INNER_PRODUCT;
         int dimension = 16;
         String modelId = "test-model-id";